#274: PR adaptations - move question observation utils into correct folder (package location)
